fix: surface background tasks after promotion
