+++ rust/crates/astra-turn-core/src/lib.rs
@@ -14,7 +14,7 @@
 pub mod action_compensation;
 pub mod activity;
 pub mod agent_live_event;
-pub mod agent_tree_state;
+
 pub mod agentic;
 pub mod alert_dispatcher;
 pub mod approval;
